Places components in separate file, removes old checkbox implementation
